results/run-20230924-033218_LLMatic-zero_shot_donnemartin-system-design-oop-whatsapp-donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_5-runs_log/dropbox/1/LLMatic/cloudsafe/tests/test_user_service.py
import pytest
from cloudsafe.app.models import User
from cloudsafe.app.user_service import UserService


def test_user_service():
	user_service = UserService()

	# Register a user
	assert user_service.register_user(1, 'John Doe', 'john@example.com', 'password', 'profile.jpg') == 'User registered successfully'

	# Login the user
	assert user_service.login_user('john@example.com', 'password') == 'Login successful'

	# Change the user's password
	assert user_service.change_password('john@example.com', 'password', 'new_password') == 'Password changed successfully'

	# Update the user's profile picture
	assert user_service.update_profile_picture('john@example.com', 'new_profile.jpg') == 'Profile picture updated successfully'
